test(records): the slot mirror lives beside the helper it mirrors
The root package may not import a module (ADR-0054), and a mirror has to
CALL the function rather than read it as text — so the census stays in
backendarch, where it needs no import, and the mirror moves into people.
